pyproject.toml
[tool.poetry]
name = "swpt_payments"
version = "0.1.0"
description = "Swaptacular micro-service that manages payments"
authors = ["Evgeni Pandurski <epandurski@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.5"
pudb = "*"
web-pdb = "*"
gunicorn = "*"
python-json-logger = "^0.1.11"
flask-env = "^2.0"
flask-migrate = "^2.4"
flask_signalbus = "^0.4.0"
flask_melodramatiq = "^0.3.4"
psycopg2-binary = "^2.8"
dramatiq = {git = "https://github.com/epandurski/dramatiq.git", extras = ["rabbitmq", "watch"], branch = "set-queue-name-if-missing"}
pytest = "^4.0"
pytest-mock = "^1.10"

[tool.poetry.dev-dependencies]
python-dotenv = ">=0.10.1"
flake8 = "^3.7"
pytest-pudb = "^0.7.0"
pytest-dotenv = "^0.4.0"
pytest-cov = "^2.7"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
